requirements.txt
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
--extra-index-url https://download.pytorch.org/whl/cu121
-f https://www.paddlepaddle.org.cn/whl/windows/mkl/avx/stable.html
fastapi
uvicorn
pydantic
python-multipart
llama-cpp-python==0.3.4
pytest
numpy==1.26.4
opencv-python==4.6.0.66
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
ultralytics
paddleocr==2.7.0.3
paddlepaddle-gpu==2.6.2
shapely
pyclipper
Pillow
